requirements.txt
anthropic>=0.39.0
openai>=1.50.0
google-generativeai>=0.8.0
httpx>=0.27.0
tabulate>=0.9.0
